+++ devel/CHANGELOG.md
@@ -10,11 +10,15 @@
 - [trvl]: trivial modifications;
 - [warn]: fixing compiler warning messages;
 
-## 2021-08-07:
-- XDHBRD: fixing connect/broadcast conflicts,
+## 2021-08-14:
+- XDHBRD: [rvrt] 2021-08-07;
+
+## 2021-08-11:
+- XDHCMN: introducing *Dismis* special action,
 
 ## 2021-08-07:
 - IDXBTQ: [adpt]
+- XDHBRD: fixing connect/broadcast conflicts,
 
 ## 2021-06-30: